Back/quartzo/requirements-dev.txt
asgiref==3.4.1
autopep8
black
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-app-novadata
django-avatar
django-browser-reload
django-filter
django-import-export
django-widget-tweaks
django_debug_toolbar
djangorestframework
djangorestframework-simplejwt
docutils
drf-spectacular
drf-spectacular-sidecar
flake8
Markdown
mypy
mypy-extensions
novadata-utils
pyflakes
python-decouple==3.6
pytz==2021.3
setuptools
sqlparse==0.4.2
static3==0.7.0
twine
wheel
django-advanced-filters